patches/algodatta-hotfix/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
